Permit decimal billing item values

+++ cloudesire-api-client-dto/src/main/java/com/cloudesire/platform/apiclient/dto/model/dto/SubscriptionV1DTO.java
@@ -1,25 +1,26 @@
 package com.cloudesire.platform.apiclient.dto.model.dto;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
 public class SubscriptionV1DTO extends BaseSubscriptionDTO
 {
-    private Map<UrlEntityDTO, Integer> billingItems = new HashMap<>();
+    private Map<UrlEntityDTO, BigDecimal> billingItems = new HashMap<>();
 
-    public Map<UrlEntityDTO, Integer> getBillingItems()
+    public Map<UrlEntityDTO, BigDecimal> getBillingItems()
     {
         return billingItems;
     }
 
-    public SubscriptionV1DTO setBillingItems( Map<String, Integer> billingItems )
+    public SubscriptionV1DTO setBillingItems( Map<String, BigDecimal> billingItems )
     {
         if ( billingItems == null ) return this;
 
         this.billingItems = new HashMap<>();
 
-        for ( Map.Entry<String, Integer> entry : billingItems.entrySet() )
+        for ( Map.Entry<String, BigDecimal> entry : billingItems.entrySet() )
         {
             this.billingItems.put( new UrlEntityDTO( entry.getKey() ), entry.getValue() );
         }
